// src/app/page/master/golongan-asuransi/golongan-asuransi.interface.ts
import {EventEmitter} from '@angular/core';

export interface GolonganAsuransi {
    kodeGolonganDarah?: string;
	namaGolonganDarah?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	id?: any;
	kdProfile?: string;
	kode?: any;
	version?: any;
	label?: any;

}